backend/pyproject.toml
[tool.poetry]
name = "talia-backend"
version = "0.1.0"
description = "Backend FastAPI modular para TalIA"
authors = ["TalIA Team <hola@talia.mx>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic-settings = "^2.2"
httpx = "^0.27.0"
openai = "^1.30.0"
twilio = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
ruff = "^0.4.1"
pre-commit = "^3.7.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:Please use `import python_multipart`:PendingDeprecationWarning",
]

[tool.ruff]
line-length = 100
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
